fix: silence compiler warnings

+++ util.c
@@ -1,7 +1,7 @@
-
-#include "util.h"
 #include <stdio.h>
 #include <string.h>
+
+#include "util.h"
 
 int clamp_int(int a, int min, int max)
 {
@@ -12,7 +12,7 @@ int clamp_int(int a, int min, int max)
 
 bool str_starts_with(const char* substr, const char* str)
 {
-    for (int i = 0; i < strlen(substr); ++i) {
+    for (size_t i = 0; i < strlen(substr); ++i) {
         if (str[i] != substr[i]) {
             return false;
         }
@@ -22,7 +22,7 @@ bool str_starts_with(const char* substr, const char* str)
 
 bool str_ends_with(const char* substr, const char* str)
 {
-    for (int i = 0; i < strlen(substr); ++i) {
+    for (size_t i = 0; i < strlen(substr); ++i) {
         if (str[strlen(str)-1-i] != substr[strlen(substr)-1-i]) {
             return false;
         }
